requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
python-dotenv==0.19.2
gunicorn==20.1.0
